cmd/libsnap-confine-private: fix build on 32bit systems
On Ubuntu 18.04 32bit:

libsnap-confine-private/bpf-support.c: In function ‘bpf_is_bpffs’:
libsnap-confine-private/bpf-support.c:182:19: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
     if (fs.f_type == BPF_FS_MAGIC) {
                   ^~
cc1: all warnings being treated as errors

The manpage of statfs(2) mentions the confusion about signedness of the field,
suggesting to use unsigned int when working with variables. Yet somehow almost
identical code in sc_group_is_v2() in cgroup-support.c compiles without hiccups.

+++ cmd/libsnap-confine-private/bpf-support.c
@@ -178,7 +178,8 @@ bool bpf_path_is_bpffs(const char *path) {
         }
         die("cannot check filesystem type of %s", path);
     }
-    if (fs.f_type == BPF_FS_MAGIC) {
+    /* see statfs(2) notes on  __fsword_t */
+    if ((unsigned int)fs.f_type == BPF_FS_MAGIC) {
         return true;
     }
     return false;
